move error logging to its own module
